refactor(wingman): extract WingmanSkillManager service
Move skill discovery, init, enable/disable, and prepare/unprepare
logic out of the unified Wingman class. Dedupe the ~50 lines of
yaml-loading + override-merging + platform-checking shared between
init_skills and enable_skill into private helpers.
